Move builder to from builder to interface.
