+++ core/src/main/java/org/stellar/anchor/sep31/RefundPayment.java
@@ -25,4 +25,25 @@ public interface RefundPayment {
         .refundedAt(null)
         .build();
   }
+
+  /**
+   * Will create a SEP-31 RefundPayment object out of a PlatformApi RefundPayment object.
+   *
+   * @param platformApiRefundPayment is the platformApi's RefundPayment object.
+   * @param factory is a Sep31TransactionStore instance used to build the object.
+   * @return a SEP-31 RefundPayment object.
+   */
+  static RefundPayment of(
+      org.stellar.anchor.api.shared.RefundPayment platformApiRefundPayment,
+      Sep31TransactionStore factory) {
+    if (platformApiRefundPayment == null) {
+      return null;
+    }
+
+    RefundPayment refundPayment = factory.newRefundPayment();
+    refundPayment.setId(platformApiRefundPayment.getId());
+    refundPayment.setAmount(platformApiRefundPayment.getAmount().getAmount());
+    refundPayment.setFee(platformApiRefundPayment.getFee().getAmount());
+    return refundPayment;
+  }
 }
