+++ core/src/main/java/org/stellar/anchor/sep31/Refunds.java
@@ -1,6 +1,9 @@
 package org.stellar.anchor.sep31;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 import org.stellar.anchor.api.shared.Amount;
 
 public interface Refunds {
@@ -34,5 +37,31 @@ public interface Refunds {
         .amountFee(new Amount(getAmountFee(), assetName))
         .payments(payments)
         .build();
+  }
+
+  /**
+   * Will create a SEP-31 Refunds object out of a PlatformApi Refund object.
+   *
+   * @param platformApiRefunds is the platformApi's Refund object.
+   * @param factory is a Sep31TransactionStore instance used to build the object.
+   * @return a SEP-31 Refunds.
+   */
+  static Refunds of(
+      org.stellar.anchor.api.shared.Refund platformApiRefunds, Sep31TransactionStore factory) {
+    if (platformApiRefunds == null) {
+      return null;
+    }
+
+    Refunds refunds = factory.newRefunds();
+    refunds.setAmountRefunded(platformApiRefunds.getAmountRefunded().getAmount());
+    refunds.setAmountFee(platformApiRefunds.getAmountFee().getAmount());
+
+    ArrayList<RefundPayment> payments =
+        Arrays.stream(platformApiRefunds.getPayments())
+            .map(refundPayment -> RefundPayment.of(refundPayment, factory))
+            .collect(Collectors.toCollection(ArrayList::new));
+    refunds.setRefundPayments(payments);
+
+    return refunds;
   }
 }
